requirements.txt
fastapi==0.95.2
pydantic==1.10.7
cassandra-driver==3.25.0
uvicorn==0.22.0
robotframework
python-multipart
pyjwt
passlib[bcrypt]